+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to
 - cpp: add std/unique to test std::unique();
 - kernel: add version.sh;
 #### Changed
+- gprof: move to gnu/binutils/ dir;
 - `make_tl`: only use `make_tl` for user `rongtao`;
 #### Deprecated
 #### Removed